Log-Threat-Detector/requirements.txt
fpdf2
ttkbootstrap
matplotlib
requests
psutil
urllib3
beautifulsoup4
fonttools
pillow
